fix: Complete staging environment post-merge cleanup
This commit properly implements all the architectural fixes after merge issues:

## Files Created/Fixed:
- **lib/api-environment.ts**: Robust API route environment detection with multiple fallback methods
- **lib/staging-service.ts**: Staging data management service (moved from scripts for better architecture)
- **Updated app/api/staging/reset/route.ts**: Now imports from lib/staging-service and uses isApiStaging()
- **Updated scripts/init-staging-data.ts**: Now just a CLI wrapper calling the service
- **Updated middleware.ts**: Fixed matcher to include API routes (removed 'api' exclusion)
- **Updated lib/mock-users.ts**: Exported BASE_MOCK_USERS for service reuse

## Architecture Improvements:
- ✅ Clean separation: scripts (CLI) vs lib (business logic)
- ✅ API routes use specialized environment detection
- ✅ Middleware properly handles /staging/api/* routes
- ✅ Single source of truth for mock user data
- ✅ Proper error handling and service patterns

Build successful - all staging environment functionality now working correctly.
